refactor needed on adapter, entities and fragments
