+++ app/src/main/java/com/globant/pokemontcgapp/fragment/PokemonSupertypeFragment.kt
@@ -8,8 +8,8 @@ import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.GridLayoutManager
 import com.globant.domain.entity.PokemonSupertype
-import com.globant.pokemontcgapp.adapter.PokemonSupertypesAdapter
-import com.globant.pokemontcgapp.databinding.FragmentPokemonSupertypeLayoutBinding
+import com.globant.pokemontcgapp.adapter.PokemonSecondaryTypesAdapter
+import com.globant.pokemontcgapp.databinding.FragmentPokemonAlltypesLayoutBinding
 import com.globant.pokemontcgapp.util.Event
 import com.globant.pokemontcgapp.util.getColumnsByOrientation
 import com.globant.pokemontcgapp.util.pokemonSupertypesResources
@@ -21,10 +21,10 @@ import org.koin.androidx.viewmodel.ext.android.viewModel
 class PokemonSupertypeFragment : Fragment() {
 
     private val pokemonSupertypeViewModel by viewModel<PokemonSupertypeViewModel>()
-    private lateinit var binding: FragmentPokemonSupertypeLayoutBinding
+    private lateinit var binding: FragmentPokemonAlltypesLayoutBinding
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        binding = FragmentPokemonSupertypeLayoutBinding.inflate(inflater, container, false)
+        binding = FragmentPokemonAlltypesLayoutBinding.inflate(inflater, container, false)
         return binding.root
     }
 
@@ -36,22 +36,22 @@ class PokemonSupertypeFragment : Fragment() {
 
     override fun onResume() {
         super.onResume()
-        binding.pokemonSupertypeLoading.visibility = View.GONE
+        binding.pokemonAlltypesLoading.visibility = View.GONE
         pokemonSupertypeViewModel.getPokemonSupertypes(pokemonSupertypesResources)
     }
 
     private fun updateUI(data: Event<Data<List<PokemonSupertype>>>) {
         val pokemonSupertypesData = data.getContentIfNotHandled()
         when (pokemonSupertypesData?.status) {
-            Status.LOADING -> binding.pokemonSupertypeLoading.visibility = View.VISIBLE
+            Status.LOADING -> binding.pokemonAlltypesLoading.visibility = View.VISIBLE
             Status.SUCCESS -> pokemonSupertypesData.data?.let { showPokemonSupertypes(it) }
             Status.ERROR -> showPokemonSupertypesError(pokemonSupertypesData.error?.message)
         }
     }
 
     private fun showPokemonSupertypes(pokemonSupertypes: List<PokemonSupertype>) {
-        binding.pokemonSupertypeLoading.visibility = View.GONE
-        binding.pokemonSupertypeRecyclerView.apply {
+        binding.pokemonAlltypesLoading.visibility = View.GONE
+        binding.pokemonAlltypesRecyclerView.apply {
             layoutManager =
                 GridLayoutManager(
                     context,
@@ -60,13 +60,13 @@ class PokemonSupertypeFragment : Fragment() {
                         COLUMNS_LANDSCAPE
                     )
                 )
-            adapter = PokemonSupertypesAdapter(pokemonSupertypes)
+            adapter = PokemonSecondaryTypesAdapter(pokemonSupertypes)
             visibility = View.VISIBLE
         }
     }
 
     private fun showPokemonSupertypesError(error: String?) {
-        binding.pokemonSupertypeLoading.visibility = View.GONE
+        binding.pokemonAlltypesLoading.visibility = View.GONE
         error?.let { Toast.makeText(context, it, Toast.LENGTH_SHORT).show() }
     }
