+++ domain/src/main/java/com/globant/domain/entity/SecondaryTypes.kt
@@ -1,0 +1,6 @@
+package com.globant.domain.entity
+
+open class SecondaryTypes(
+    var name: String,
+    var bgColor: Int
+)
